--- BackEnd/src/main/java/MusicAPI/virtuouso/models/RandomForest.java
@@ -1,5 +1,0 @@
-package MusicAPI.virtuouso.models;
-
-class RandomForest {
-    //TODO implement...
-}
