+++ BackEnd/src/main/java/MusicAPI/virtuouso/models/decisionTree/RandomForest.java
@@ -1,0 +1,5 @@
+package MusicAPI.virtuouso.models.decisionTree;
+
+class RandomForest {
+    //TODO implement...
+}
