+++ BackEnd/src/main/java/MusicAPI/virtuouso/models/genetic/GeneticAlgorithm.java
@@ -2,6 +2,7 @@ package MusicAPI.virtuouso.models.genetic;
 
 import java.util.*;
 
+import MusicAPI.virtuouso.models.genetic.fitness.FitnessFunction;
 import org.jenetics.*;
 import org.jenetics.engine.Engine;
